Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/rendering/non-replaced-elements/tables/table-ua-stylesheet-expected.txt


PASS Computed 'display' on table should match html spec
FAIL Computed 'box-sizing' on table should match html spec assert_equals: expected "border-box" but got "content-box"
PASS Computed 'border-spacing' on table should match html spec
PASS Computed 'border-collapse' on table should match html spec
FAIL Computed 'text-indent' on table should match html spec assert_equals: expected "0px" but got "100px"
